Metrics Collection Refactoring (1)
Move apksize and coverage gradle tasks defination and plugins to different packages.
