+++ buildSrc/src/main/groovy/com/google/firebase/gradle/plugins/measurement/apksize/GenerateMeasurementsTask.groovy
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 
-package com.google.firebase.gradle.plugins.measurement
+package com.google.firebase.gradle.plugins.measurement.apksize
 
 import org.gradle.api.DefaultTask
 import org.gradle.api.Task
