+++ buildSrc/src/test/groovy/com/google/firebase/gradle/plugins/measurement/apksize/ApkSizeTestProject.groovy
@@ -4,7 +4,7 @@
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//            http://www.apache.org/licenses/LICENSE-2.0
+//      http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -12,7 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.firebase.gradle.plugins.measurement
+package com.google.firebase.gradle.plugins.measurement.apksize
+
+import com.android.utils.FileUtils
 
 import java.nio.file.Files
 import java.nio.file.Path
@@ -26,7 +28,7 @@ import org.junit.rules.TemporaryFolder
 class ApkSizeTestProject extends ExternalResource {
 
     private static final String BUILD_GRADLE = """
-        import com.google.firebase.gradle.plugins.measurement.GenerateMeasurementsTask
+        import com.google.firebase.gradle.plugins.measurement.apksize.GenerateMeasurementsTask
         import com.google.firebase.gradle.plugins.measurement.UploadMeasurementsTask
 
         buildscript {
@@ -154,10 +156,7 @@ class ApkSizeTestProject extends ExternalResource {
         def dest = buildSrc.resolve(src)
 
         Files.createDirectories(dest)
-        Files.list(src)
-            .map { path -> path.getFileName().toString() }
-            .filter { name -> name.endsWith(".groovy") }
-            .forEach { name -> Files.copy(src.resolve(name), dest.resolve(name)) }
+        FileUtils.copyDirectory(src.toFile(), dest.toFile())
     }
 
     /** Creates the fake project files for the temporary project. */
