+++ subprojects.cfg
@@ -16,5 +16,6 @@ transport:transport-runtime
 transport:transport-backend-cct
 
 
-tools:apksize
+tools:measurement:apksize
+tools:measurement:coverage
 tools:errorprone
