+++ tools/measurement/apksize/debug.keystore
