+++ tools/measurement/apksize/proguard-rules-aggressive.pro
