+++ tools/measurement/apksize/src/database/database.gradle
