+++ tools/measurement/apksize/src/empty/empty.gradle
