+++ tools/measurement/apksize/src/firestore/firestore.gradle
