+++ tools/measurement/apksize/src/functions/functions.gradle
