+++ tools/measurement/apksize/src/main/AndroidManifest.xml
