+++ tools/measurement/apksize/src/main/java/com.google.apksize/SampleCode.java
