+++ tools/measurement/apksize/src/main/res/values/colors.xml
