+++ tools/measurement/apksize/src/main/res/values/strings.xml
