+++ tools/measurement/apksize/src/storage/java/com.google.apksize/Storage.java
