+++ tools/measurement/apksize/src/storage/res/values/strings.xml
